rename source directory
Without plans to transpile the source, it makes sense to rename
the souce directory back to library. Negligible change, but at
the end of the day... what're you doin' looking through my commit
history! Get outta here! Scram!
