+++ karma.conf.js
@@ -34,7 +34,7 @@ module.exports = function (config) {
 		plugins: [
 			'karma-jasmine',
 			'karma-phantomjs-launcher',
-			require('./src'),
+			require('./lib'),
 		],
