+++ package.json
@@ -2,7 +2,7 @@
   "name": "karma-rollup-preprocessor",
   "version": "2.0.2",
   "description": "Karma preprocessor to bundle ES6 modules using Rollup",
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "scripts": {
     "test": "./node_modules/karma/bin/karma start karma.conf.js"
   },
